Moved the tests for lint `search_is_some` to new files `search_is_some.rs` and `search_is_some_fixable.rs`
